Add files via upload
The main difference between our final version and the first version is the addition of the store class. This contained a lot of the important information such as the inventory, the register, etc... and tracked those things. A couple attributes were also changed as we realized we needed or didn't need them. The staff interface was also changed into an abstract class in the actual implementation of our code. The customers methods were also changed to simplify the code.