Add files via upload
A big difference between this UML and the original one is that this one includes the Owner class in place of the User class from the original. This owner class owns the different stores and allows for command input in the console. Another change is the StaffPool now being implemented through an interface. A change to the Abstract Factory portion is that the guitar parts now all have the superclass Parts which they inherit a price and method from. Other then a few other small changes the UML remained fairly similar to how it was originally designed.